After a great deal of messing around, I got the abstract working properly
Note that this NEEDS the most recent (as in, from the last 20 days) version of rmarkdown, for the compact-title option to work. I had to enable this, as by default Rmarkdown is enabling the titling package, which messes with the PeerJ style sheet, and was preventing the abstract from appearing.