Fixing a couple other compilation bugs.
Had to add explicit figure labels back in to captions. For some reason, Rmarkdown was automatically adding labels to some figures, but not to others; to be on the safe side, I've added labels to all figures. Also fixed a couple places where we should be refering to figures as a, b, c, etc. rather than left and right.